+++ app.py
@@ -1,21 +1,15 @@
+import os
 import streamlit as st
-import pandas as pd
+from pages import demo, alpaca_live
 
+
+st.set_page_config(layout="wide")
 st.title("Stock price vs time")
 
-# load demo data to a dataframe
-df = pd.read_csv('./resources/demo_stock_data.csv', parse_dates=["ts"])
-
-# get unique symbols
-symbols = df["symbol"].unique().tolist()
-
-# sidemenu
 st.sidebar.header("Menu")
-symbolSelected = st.sidebar.selectbox("Symbol", symbols, index=0)
+page = st.sidebar.selectbox("Page", ["Demo Data", "Alpaca Live"]) 
 
-# filter dataframe by selected symbol and sort by timestamp
-df_sym = df[df["symbol"] == symbolSelected].sort_values("ts").set_index("ts")
-
-# line chart of price vs timestamp
-st.subheader(f"{symbolSelected} price over time")
-st.line_chart(df_sym["price"])
+if page == "Demo Data":
+    demo.demo_page()
+else:
+    alpaca_live.alpaca_page()